+++ MatchRetriever/Enumerals/SubscriptionType.cs
@@ -8,5 +8,10 @@ namespace MatchRetriever.Enumerals
         Free = 1,
         Premium = 2,
         Ultimate = 3,
+
+        /// <summary>
+        /// Users who have MENTOR.GG in their name.
+        /// </summary>
+        Influencer = 4
     }
 }