+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.3.1] - 2020-06-17
+## [2.4.0] - 2020-06-19
 ### Added
 - `forceRefresh` param to PlayerInfo endpoint
-- New endpoint match/{matchId}/players 
+- New endpoint match/{matchId}/players
+- Support for SubscriptionType Influencer
 
 ## [2.2.0] - 2020-06-08
 ### Added